third_party/blink/web_tests/fast/forms/calendar-picker/calendar-picker-pre-100-year.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/js-test.js"></script>
<script src="../../../resources/testdriver.js"></script>
<script src="../../../resources/testdriver-vendor.js"></script>
<script src="../../forms/resources/common.js"></script>
<script src="../../forms/resources/picker-common.js"></script>
<script src="resources/calendar-picker-common.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<input type=date id=date value="0002-02-02">
<script>
description('Check if year before 100 can be set.');

openPickerDeprecatedJsTest(document.getElementById('date'), finishTest);

function finishTest() {
    var daysContainer = internals.pagePopupWindow.document.getElementsByClassName("days-area")[0];
    eventSender.keyDown('ArrowRight');
    removeCommitDelay();
    eventSender.keyDown('Enter');
    shouldBeEqualToString('document.getElementById("date").value', '0002-02-03');
    finishJSTest();
}
</script>
</body>
</html>
